Conflicts:
	src/ClassAdminBackEnd/EntityType.java
	src/ClassAdminBackEnd/Format.java
	src/ClassAdminBackEnd/Project.java

+++ src/ClassAdminBackEnd/MarkEntity.java
@@ -39,8 +39,8 @@ public class MarkEntity extends SuperEntity{
 		return this.getMark();
 	}
 
-	public int saveToDB(SqlJetDb db, int parentID, PDatIDGenerator idgen) throws SqlJetException {
-		int id = super.saveToDB(db, parentID, idgen);
+	public long saveToDB(SqlJetDb db, long parentID, PDatIDGenerator idgen) throws SqlJetException {
+		long id = super.saveToDB(db, parentID, idgen);
 		db.beginTransaction(SqlJetTransactionMode.WRITE);
         try {
         	//TODO
